+++ ChangeLog.md
@@ -1,5 +1,6 @@
 ## [UNRELEASED]
 - Re-connect when losing connection to etcd
+- Support authentication against etcd
 
 ## [1.0.0] 2024-10-15
 - Added basic policy and authentication support